api/requirements.txt
azure-functions
motor==3.3.1
pymongo==4.5.0
python-dotenv>=1.0.1